requirements.txt
Flask==3.0.0
Pillow==10.2.0
ImageHash==4.3.1
gunicorn==22.0.0
